feat: embargo release version
This little change means we can work on preparing the stable release
without it appearing on the Keyman websites, which allows us to deal
with broken builds, etc, and not cause chaos.

This is particularly helpful when the website changes themselves are
linked to the release version, which is certainly the case for the 14.0
release.