fix: improve history line matches for platforms
Fixes #44.

Tightens the regex matching for history entries to exclude references
to platform in the message body of a history entry.

Also loosens the regex matches for specific scopes so we match any scope
starting with the platform name.

Finally, updates to the new scopes introduced in 15.0 (e.g. `core`)